JsonBadPlanTests: move runtime-only errors back to systemtests Cannot be caught until server is running.
